+++ src/assets/images/prev_icon.png
