+++ src/components/setting/EnvCharacterList.tsx
@@ -4,12 +4,12 @@ import CharacterPreview from "./CharacterPreview";
 import OptionItem from "components/studySetting/OptionItem";
 import { Character } from "components/studySetting/Character";
 import { Voice } from "components/studySetting/Voice";
-import ImgBox from "../../assets/Img_box.png";
+import ImgBox from "../../assets/images/Img_box.png";
 
 {/* 임시 character 임시 더미 데이터 */}
 const characterData: Character[] = [
-  { id: 1, name: 'noti', img: require('../../assets/Noti.png'), large_img: require('../../assets/Noti2.png') },
-  { id: 2, name: 'woman', img: require('../../assets/avatar_woman.png'), large_img: require('../../assets/avatar_woman.png') },
+  { id: 1, name: 'noti', img: require('../../assets/images/Noti.png'), large_img: require('../../assets/images/Noti2.png') },
+  { id: 2, name: 'woman', img: require('../../assets/images/avatar_woman.png'), large_img: require('../../assets/images/avatar_woman.png') },
 ];
 
 const EnvCharacterList: React.FC = () => {
